requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.35
alembic==1.13.2
apscheduler==3.10.4
